requirements.txt
fastapi
uvicorn
requests
web3
pydantic
python-dotenv